+++ web-app/src/App.tsx
@@ -1,14 +1,16 @@
 // App.tsx
 import NavBar from './components/NavBar';
 import Header from './components/Header';
-import './components/Styling/Global.css';
-import './components/Styling/theme.css';
+import './Styling/Global.css';
+import './Styling/theme.css';
+import { Outlet } from 'react-router-dom';
 
 const App = () => {
   return (
     <div className="app">
       <Header />
       <NavBar />
+      <Outlet />
     </div>
   );
 };
