+++ web-app/src/Styling/NavBar.css
