+++ web-app/src/Styling/PantryView.css
