+++ web-app/src/components/RecipeCarousel.tsx
@@ -4,7 +4,7 @@ import axios from 'axios';
 import Slider from 'react-slick';
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
-import './Styling/RecipeCarousel.css'; // Ensure your CSS file path is correct
+import '../Styling/RecipeCarousel.css'; // Ensure your CSS file path is correct
 
 interface Recipe {
     _id: string;
